pyproject.toml
[tool.poetry]
name = "candfans-post-analyzer"
version = "0.1.0"
description = ""
authors = ["komanushi <155377273+komanushi@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
psycopg = "^3.1.18"
candfans-client = "^3.0.0"
pydantic = "^2.6.3"
factory-boy = "^3.3.0"
django-extensions = "^3.2.3"
ipython = "^8.22.1"
redis = "^5.0.2"
hiredis = "^2.3.2"
django-rq = "^2.10.1"
watchdog = "^4.0.0"
freezegun = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
